// apps/web-staking/src/app/components/createPool/enums/enums.ts
export enum ERROR_MESSAGE {
  ENTER_TEXT = "You must enter a name for your pool",
  NAME_LENGTH = "Pool name must be less than 24 characters",
  DESCRIPTION = "You must enter a description for your pool",
  DESCRIPTION_LENGTH = "Pool description must be less than 400 characters",
  BAD_WORD_MESSAGE = "Please refrain of using any banned words like ",
  TRACKER_NAME = "Tracker name must be less than 24 characters",
  TICKER = "Ticker must be less than 6 characters",
}

export enum LABELS {
  NAME = "Pool name*",
  DESCRIPTION = "Pool description*",
}

export enum PLACEHOLDERS {
  NAME = "Enter pool name here",
  DESCRIPTION = "Enter pool description here",
}